+++ examples/oneapi/json_scraper_oneapi.py
@@ -3,10 +3,8 @@ Basic example of scraping pipeline using JSONScraperGraph from JSON documents
 """
 
 import os
-from dotenv import load_dotenv
 from scrapegraphai.graphs import JSONScraperGraph
 from scrapegraphai.utils import convert_to_csv, convert_to_json, prettify_exec_info
-load_dotenv()
 
 # ************************************************
 # Read the JSON file
@@ -22,8 +20,6 @@ with open(file_path, 'r', encoding="utf-8") as file:
 # ************************************************
 # Define the configuration for the graph
 # ************************************************
-
-openai_key = os.getenv("OPENAI_APIKEY")
 
 graph_config = {
     "llm": {