+++ examples/oneapi/search_graph_oneapi.py
@@ -2,11 +2,8 @@
 Example of Search Graph
 """
 
-import os
-from dotenv import load_dotenv
 from scrapegraphai.graphs import SearchGraph
 from scrapegraphai.utils import convert_to_csv, convert_to_json, prettify_exec_info
-load_dotenv()
 
 # ************************************************
 # Define the configuration for the graph