+++ scrapegraphai/nodes/generate_answer_csv_node.py
@@ -49,7 +49,7 @@ class GenerateAnswerCSVNode(BaseNode):
         input: str,
         output: List[str],
         node_config: Optional[dict] = None,
-        node_name: str = "GenerateAnswer",
+        node_name: str = "GenerateAnswerCSV",
     ):
         """
         Initializes the GenerateAnswerNodeCsv with a language model client and a node name.
@@ -58,6 +58,7 @@ class GenerateAnswerCSVNode(BaseNode):
             node_name (str): name of the node
         """
         super().__init__(node_name, "node", input, output, 2, node_config)
+        
         self.llm_model = node_config["llm_model"]
         self.verbose = (
             False if node_config is None else node_config.get("verbose", False)
@@ -93,7 +94,12 @@ class GenerateAnswerCSVNode(BaseNode):
         user_prompt = input_data[0]
         doc = input_data[1]
 
-        output_parser = JsonOutputParser()
+        # Initialize the output parser
+        if self.node_config.get("schema", None) is not None:
+            output_parser = JsonOutputParser(pydantic_object=self.node_config["schema"])
+        else:
+            output_parser = JsonOutputParser()
+
         format_instructions = output_parser.get_format_instructions()
    
         chains_dict = {}